[PT FE] Separate scripting and tracing in decoder
