+++ src/bindings/python/src/openvino/frontend/pytorch/ts_decoder.py
@@ -7,29 +7,16 @@
 from openvino.frontend.pytorch.py_pytorch_frontend import _FrontEndPytorchDecoder as Decoder
 from openvino.frontend.pytorch.py_pytorch_frontend import _Type as DecoderType
 from openvino.runtime import op, PartialShape, Type as OVType, OVAny
-from openvino.frontend.pytorch.utils import ivalue_to_constant, get_value_from_getattr, pt_to_ov_type_map, torch_tensor_to_ov_const
+from openvino.frontend.pytorch.utils import ivalue_to_constant, get_value_from_getattr, pt_to_ov_type_map, prepare_example_inputs_and_model, convert_quantized_tensor
 from openvino.runtime import opset11 as ops
 
 import typing
 import torch
 import numpy as np
 
-wrapper_template = """
-import torch
-from typing import *
-
-class ModelWrapper(torch.nn.Module):
-    def __init__(self, model):
-        super().__init__()
-        self.model = model
-
-    def forward(self, {input_sign}):
-        return self.model({example_input})
-"""
-
 
 class TorchScriptPythonDecoder (Decoder):
-    def __init__(self, pt_module, graph_element=None, example_input=None, alias_db=None, shared_memory=True):
+    def __init__(self, pt_module, graph_element=None, example_input=None, alias_db=None, shared_memory=True, skip_freeze=False):
         Decoder.__init__(self)
         # We store every decoder created by this decoder so that all them are not deleted until the first decoder is deleted
         self.m_decoders = []
@@ -38,11 +25,13 @@ class TorchScriptPythonDecoder (Decoder):
         self._input_is_list = False
         if graph_element is None:
             try:
-                pt_module = self._get_scripted_model(pt_module, example_input)
+                pt_module = self._get_scripted_model(pt_module, example_input, skip_freeze)
             except Exception as e:
                 if example_input is not None:
-                    msg = "tracing or scripting"
-                    help_msg = ""
+                    msg = "tracing"
+                    help_msg = "Please check correctness of provided 'example_input'. "
+                    "Sometimes models can be converted in scripted mode, please try running "
+                    "conversion without 'example_input'."
                 else:
                     msg = "scripting"
                     help_msg = "\nTracing sometimes provide better results, please provide valid 'example_input' argument. "
@@ -82,73 +71,9 @@ class TorchScriptPythonDecoder (Decoder):
                     preserved_attributes.append(name)
         return preserved_attributes
 
-    def _get_scripted_model(self, pt_module, example_inputs=None):
+    def _get_scripted_model(self, pt_module, example_inputs=None, skip_freeze=False):
         import torch
         import inspect
-
-        def process_dict_inputs(inputs, input_params, model):
-            ordered_inputs = []
-            for input_name in input_params:
-                if input_name in inputs:
-                    ordered_inputs.append(input_name)
-
-            input_signature = list(input_params)
-            if ordered_inputs == input_signature[:len(ordered_inputs)]:
-                example_inputs = [inputs[input_name] for input_name in ordered_inputs]
-                if all([isinstance(inp, torch.Tensor) for inp in example_inputs]):
-                    return {"example_inputs": [inputs[name] for name in ordered_inputs]}, ordered_inputs, model
-                return {"example_inputs": example_inputs}, ordered_inputs, model
-
-            # PyTorch has some difficulties to trace models with named unordered parameters:
-            # torch < 2.0.0 supports only positional arguments for tracing
-            # pytorch == 2.0.0 supports input kwargs tracing,
-            # but does not support complex nested objects (e. g. tuple of tuples of tensors)
-            # We will use wrapper for making them positional as workaround.
-
-            input_sign_str = []
-            input_params_str = []
-
-            for input_name in ordered_inputs:
-                if str(input_params[input_name].annotation).startswith("typing.Union"):
-                    filter_custom_args = []
-                    for arg in input_params[input_name].annotation.__args__:
-                        str_arg = str(arg)
-                        is_typing = str_arg.startswith("typing.")
-                        is_torch = "torch." in str_arg
-                        is_builten = str_arg in (str(int), str(float), str(type(None)))
-                        if not (is_typing or is_torch or is_builten):
-                            continue
-                        filter_custom_args.append(arg)
-                    input_params[input_name].annotation.__args__ = tuple(filter_custom_args)
-                input_sign_str.append(str(input_params[input_name]).replace("NoneType", "None"))
-                input_params_str.append(f"{input_name}={input_name}")
-
-            wrapper_class = wrapper_template.format(input_sign=', '.join(input_sign_str), example_input=', '.join(input_params_str))
-            result = {}
-            try:
-                exec(wrapper_class, result)
-
-                wrapped_model = result["ModelWrapper"](model)
-                wrapped_model.eval()
-            # if wrapping failed, it is better to return original model for avoid user confusion regarding error message
-            except Exception:
-                wrapped_model = model
-
-            return {"example_inputs": [inputs[name] for name in ordered_inputs]}, ordered_inputs, wrapped_model
-
-        def prepare_example_inputs_and_model(inputs, input_params, model):
-            input_signature = list(input_params)
-            if isinstance(inputs, dict):
-                return process_dict_inputs(inputs, input_params, model)
-            if isinstance(inputs, list) and len(inputs) == 1 and isinstance(inputs[0], torch.Tensor):
-                if "typing.List" in str(input_params[input_signature[0]].annotation):
-                    inputs = inputs[0].unsqueeze(0)
-                    self._input_is_list = True
-
-            if isinstance(inputs, torch.Tensor):
-                inputs = [inputs]
-            input_signature = input_signature[:len(inputs)]
-            return {"example_inputs": inputs}, input_signature, model
 
         if isinstance(pt_module, torch.nn.Module):
             pt_module.eval()
@@ -160,32 +85,23 @@ class TorchScriptPythonDecoder (Decoder):
             if example_inputs is None:
                 scripted = torch.jit.script(pt_module)
             else:
-                input_parameters, input_signature, pt_module = prepare_example_inputs_and_model(example_inputs, input_params, pt_module)
-                try:
-                    scripted = torch.jit.trace(pt_module, **input_parameters)
-                except Exception:
-                    try:
-                        scripted = torch.jit.script(pt_module)
-                    except Exception as se:
-                        try:
-                            scripted = torch.jit.trace(pt_module, **input_parameters, strict=False)
-                        except Exception as te:
-                            raise Exception(f"Tracing failed with exception {te}\nScripting failed with exception: {se}")
-            skip_freeze = False
-            for n in scripted.inlined_graph.nodes():
-                # TODO: switch off freezing for all traced models
-                if "quantize" in n.kind():
-                    # do not freeze quantized models
-                    skip_freeze = True
-                    break
-                elif "aten::to" in n.kind():
-                    first_input = next(n.inputs())
-                    if first_input.node().kind() == "prim::Constant":
-                        ivalue = first_input.toIValue()
-                        if isinstance(ivalue, torch.Tensor) and ivalue.dtype in [torch.bfloat16, torch.float16]:
-                            # do not freeze models with compressed constants
-                            skip_freeze = True
-                            break
+                input_parameters, input_signature, pt_module, self._input_is_list = prepare_example_inputs_and_model(example_inputs, input_params, pt_module)
+                scripted = torch.jit.trace(pt_module, **input_parameters, strict=False)
+            if not skip_freeze:
+                for n in scripted.inlined_graph.nodes():
+                    # TODO: switch off freezing for all traced models
+                    if "quantize" in n.kind():
+                        # do not freeze quantized models
+                        skip_freeze = True
+                        break
+                    elif "aten::to" in n.kind():
+                        first_input = next(n.inputs())
+                        if first_input.node().kind() == "prim::Constant":
+                            ivalue = first_input.toIValue()
+                            if isinstance(ivalue, torch.Tensor) and ivalue.dtype in [torch.bfloat16, torch.float16]:
+                                # do not freeze models with compressed constants
+                                skip_freeze = True
+                                break
             if not skip_freeze:
                 preserved_attrs = self._get_preserved_attributes(scripted)
                 f_model = torch.jit.freeze(scripted, preserved_attrs=preserved_attrs)
@@ -331,36 +247,6 @@ class TorchScriptPythonDecoder (Decoder):
             node.set_friendly_name(name)
         return node
 
-    @staticmethod
-    def convert_quantized_tensor(qtensor: torch.Tensor, shared_memory: bool):
-        # need to represent as Constant(u8) -> Convert(f32) -> Subtract(zero_point) -> Multiply (scale)
-        qscheme = qtensor.qscheme()  # torch.per_channel_affine (per_tensor)
-        if qscheme == torch.per_channel_affine:
-            int8_tensor = qtensor.int_repr()
-            scale = qtensor.q_per_channel_scales().numpy().astype(np.float32)  # (weight.q_scale() for per_tensor)
-            zero_point = qtensor.q_per_channel_zero_points().numpy().astype(np.float32)  # (weight.q_zero_point() for per_tensor)
-            axis = np.int32(qtensor.q_per_channel_axis())
-
-            new_shape = np.ones(len(int8_tensor.shape), dtype=np.int32)
-            new_shape[axis] = -1
-            zero_point_bc = np.reshape(zero_point, new_shape)
-            scale_bc = np.reshape(scale, new_shape)
-            
-            int8_const = torch_tensor_to_ov_const(int8_tensor, shared_memory=shared_memory)
-            convert = ops.convert(int8_const, np.float32)
-            sub = ops.subtract(convert, zero_point_bc)
-            return ops.multiply(sub, scale_bc).outputs()
-        elif qscheme == torch.per_tensor_affine:
-            int8_tensor = qtensor.int_repr()
-            scale = np.float32(qtensor.q_scale())
-            zero_point = np.float32(qtensor.q_zero_point())
-
-            int8_const = torch_tensor_to_ov_const(int8_tensor, shared_memory=shared_memory)
-            convert = ops.convert(int8_const, np.float32)
-            sub = ops.subtract(convert, zero_point)
-            return ops.multiply(sub, scale).outputs()
-        assert False, "Unsupported qscheme"
-
     def try_decode_get_attr(self):
         pt_value = get_value_from_getattr(self.graph_element, self.pt_module)
         assert pt_value is not None, "Couldn't retrieve value from prim::GetAttr"
@@ -368,7 +254,7 @@ class TorchScriptPythonDecoder (Decoder):
             # We assume this is __torch__.torch.classes.quantized.Conv2dPackedParamsBase or __torch__.torch.classes.quantized.LinearPackedParamsBase
             # TODO: but can be anything. Figure a better way to distinguish
             weight, bias = pt_value.unpack()
-            res = self.convert_quantized_tensor(weight, self._shared_memory)
+            res = convert_quantized_tensor(weight, self._shared_memory)
             if isinstance(bias, torch.Tensor):
                 res += ivalue_to_constant(bias)
             else:
