+++ src/bindings/python/src/openvino/frontend/pytorch/utils.py
@@ -10,9 +10,10 @@ import numpy as np
 import ctypes
 
 from openvino.runtime import op, Type as OVType, Shape, Tensor
+from openvino.runtime import opset11 as ops
 
 
-def maybe_convert_max_int(value : int):
+def maybe_convert_max_int(value: int):
     # FIXME: This is a convertion from 64-bit positive max integer value
     # to 32-bit positive max integer value. Find a better way to handle this.
     if value == torch.iinfo(torch.int64).max:
@@ -20,10 +21,12 @@ def maybe_convert_max_int(value : int):
     else:
         return value
 
+
 def make_constant(*args, **kwargs):
     return op.Constant(*args, **kwargs)
 
-def fetch_attr(self_module, target : str):
+
+def fetch_attr(self_module, target: str):
     """
     Fetch an attribute from the ``Module`` hierarchy of ``self.module``.
 
@@ -37,7 +40,8 @@ def fetch_attr(self_module, target : str):
     attr_itr = self_module
     for i, atom in enumerate(target_atoms):
         if not hasattr(attr_itr, atom):
-            raise RuntimeError(f"Node referenced nonexistent target {'.'.join(target_atoms[:i])}")
+            raise RuntimeError(
+                f"Node referenced nonexistent target {'.'.join(target_atoms[:i])}")
         attr_itr = getattr(attr_itr, atom)
     return attr_itr
 
@@ -84,6 +88,7 @@ def ivalue_to_constant(ivalue, shared_memory=True):
         return torch_tensor_to_ov_const(ivalue, shared_memory=shared_memory).outputs()
     return None
 
+
 def get_value_from_getattr(getattr_node, self_module):
     assert getattr_node.kind() == "prim::GetAttr", "Got node of kind not equal to prim::GetAttr"
     # GetAttr nodes can be nested
@@ -98,9 +103,11 @@ def get_value_from_getattr(getattr_node, self_module):
     while len(stack) > 0:
         node = stack.pop()
         attr_name = node.s("name")
-        assert hasattr(module, attr_name), f"No attribute with name \"{attr_name}\" found in module."
+        assert hasattr(
+            module, attr_name), f"No attribute with name \"{attr_name}\" found in module."
         module = getattr(module, attr_name)
     return module
+
 
 pt_to_ov_type_map = {
     "float": OVType.f32,
@@ -131,3 +138,120 @@ ov_to_c_type_map = {
     OVType.i32: ctypes.c_int,
     OVType.i64: ctypes.c_int64,
 }
+
+
+wrapper_template = """
+import torch
+from typing import *
+
+class ModelWrapper(torch.nn.Module):
+    def __init__(self, model):
+        super().__init__()
+        self.model = model
+
+    def forward(self, {input_sign}):
+        return self.model({example_input})
+"""
+
+
+def process_dict_inputs(inputs, input_params, model):
+    ordered_inputs = []
+    for input_name in input_params:
+        if input_name in inputs:
+            ordered_inputs.append(input_name)
+
+    input_signature = list(input_params)
+    if ordered_inputs == input_signature[:len(ordered_inputs)]:
+        example_inputs = [inputs[input_name] for input_name in ordered_inputs]
+        if all([isinstance(inp, torch.Tensor) for inp in example_inputs]):
+            return {"example_inputs": [inputs[name] for name in ordered_inputs]}, ordered_inputs, model
+        return {"example_inputs": example_inputs}, ordered_inputs, model
+
+    # PyTorch has some difficulties to trace models with named unordered parameters:
+    # torch < 2.0.0 supports only positional arguments for tracing
+    # pytorch == 2.0.0 supports input kwargs tracing,
+    # but does not support complex nested objects (e. g. tuple of tuples of tensors)
+    # We will use wrapper for making them positional as workaround.
+
+    input_sign_str = []
+    input_params_str = []
+
+    for input_name in ordered_inputs:
+        if str(input_params[input_name].annotation).startswith("typing.Union"):
+            filter_custom_args = []
+            for arg in input_params[input_name].annotation.__args__:
+                str_arg = str(arg)
+                is_typing = str_arg.startswith("typing.")
+                is_torch = "torch." in str_arg
+                is_builten = str_arg in (str(int), str(float), str(type(None)))
+                if not (is_typing or is_torch or is_builten):
+                    continue
+                filter_custom_args.append(arg)
+            input_params[input_name].annotation.__args__ = tuple(
+                filter_custom_args)
+        input_sign_str.append(
+            str(input_params[input_name]).replace("NoneType", "None"))
+        input_params_str.append(f"{input_name}={input_name}")
+
+    wrapper_class = wrapper_template.format(input_sign=', '.join(
+        input_sign_str), example_input=', '.join(input_params_str))
+    result = {}
+    try:
+        exec(wrapper_class, result)
+
+        wrapped_model = result["ModelWrapper"](model)
+        wrapped_model.eval()
+    # if wrapping failed, it is better to return original model for avoid user confusion regarding error message
+    except Exception:
+        wrapped_model = model
+
+    return {"example_inputs": [inputs[name] for name in ordered_inputs]}, ordered_inputs, wrapped_model
+
+
+def prepare_example_inputs_and_model(inputs, input_params, model):
+    input_is_list = False
+    input_signature = list(input_params)
+    if isinstance(inputs, dict):
+        return process_dict_inputs(inputs, input_params, model)
+    if isinstance(inputs, list) and len(inputs) == 1 and isinstance(inputs[0], torch.Tensor):
+        if "typing.List" in str(input_params[input_signature[0]].annotation):
+            inputs = inputs[0].unsqueeze(0)
+            input_is_list = True
+
+    if isinstance(inputs, torch.Tensor):
+        inputs = [inputs]
+    input_signature = input_signature[:len(inputs)]
+    return {"example_inputs": inputs}, input_signature, model, input_is_list
+
+
+def convert_quantized_tensor(qtensor: torch.Tensor, shared_memory: bool):
+    # represents torch quantized tensor as
+    # Constant(u8) -> Convert(f32) -> Subtract(zero_point) -> Multiply(scale)
+    qscheme = qtensor.qscheme()
+    if qscheme == torch.per_channel_affine:
+        int8_tensor = qtensor.int_repr()
+        scale = qtensor.q_per_channel_scales().numpy().astype(np.float32)
+        zero_point = qtensor.q_per_channel_zero_points().numpy().astype(np.float32)
+        axis = np.int32(qtensor.q_per_channel_axis())
+
+        new_shape = np.ones(len(int8_tensor.shape), dtype=np.int32)
+        new_shape[axis] = -1
+        zero_point_bc = np.reshape(zero_point, new_shape)
+        scale_bc = np.reshape(scale, new_shape)
+
+        int8_const = torch_tensor_to_ov_const(
+            int8_tensor, shared_memory=shared_memory)
+        convert = ops.convert(int8_const, np.float32)
+        sub = ops.subtract(convert, zero_point_bc)
+        return ops.multiply(sub, scale_bc).outputs()
+    elif qscheme == torch.per_tensor_affine:
+        int8_tensor = qtensor.int_repr()
+        scale = np.float32(qtensor.q_scale())
+        zero_point = np.float32(qtensor.q_zero_point())
+
+        int8_const = torch_tensor_to_ov_const(
+            int8_tensor, shared_memory=shared_memory)
+        convert = ops.convert(int8_const, np.float32)
+        sub = ops.subtract(convert, zero_point)
+        return ops.multiply(sub, scale).outputs()
+    assert False, "Unsupported qscheme"
