+++ tests/layer_tests/pytorch_tests/pytorch_layer_test_class.py
@@ -77,18 +77,16 @@ class PytorchLayerTest:
             self.torch_compile_backend_test(model, torch_inputs, custom_eps)
         else:
             with torch.no_grad():
-                model.eval()
                 trace_model = kwargs.get('trace_model', False)
                 freeze_model = kwargs.get('freeze_model', True)
-                model, converted_model = self.convert_directly_via_frontend(model, torch_inputs, trace_model, dynamic_shapes, ov_inputs, freeze_model)
-                graph = model.inlined_graph
+                smodel, converted_model = self.convert_directly_via_frontend(model, torch_inputs, trace_model, dynamic_shapes, ov_inputs, freeze_model)
 
-                if kind is not None and not isinstance(kind, (tuple, list)):
-                    kind = [kind]
-                if kind is not None:
-                    for op in kind:
-                        assert self._check_kind_exist(
-                            graph, op), f"Operation {op} type doesn't exist in provided graph"
+            if kind is not None and not isinstance(kind, (tuple, list)):
+                kind = [kind]
+            if kind is not None:
+                for op in kind:
+                    assert self._check_kind_exist(
+                        smodel.inlined_graph, op), f"Operation {op} type doesn't exist in provided graph"
             # OV infer:
             core = Core()
             compiled = core.compile_model(converted_model, ie_device)
@@ -99,7 +97,7 @@ class PytorchLayerTest:
                 return
 
             # Framework infer:
-            fw_res = model(*deepcopy(torch_inputs))
+            fw_res = smodel(*deepcopy(torch_inputs))
 
             if not isinstance(fw_res, (tuple)):
                 fw_res = (fw_res,)
@@ -182,23 +180,15 @@ class PytorchLayerTest:
         return model, om
 
     def convert_directly_via_frontend(self, model, example_input, trace_model, dynamic_shapes, ov_inputs, freeze_model):
-        import torch
-
         fe_manager = FrontEndManager()
         fe = fe_manager.load_by_framework('pytorch')
 
-        model.eval()
-        with torch.no_grad():
-            if trace_model:
-                model = torch.jit.trace(model, example_input)
-            else:
-                model = torch.jit.script(model)
-        if freeze_model:
-            _model = torch.jit.freeze(model)
+        if trace_model:
+            decoder = TorchScriptPythonDecoder(model, example_input=example_input, skip_freeze=not freeze_model)
         else:
-            _model = model
-        print(_model.inlined_graph)
-        decoder = TorchScriptPythonDecoder(_model)
+            decoder = TorchScriptPythonDecoder(model, skip_freeze=not freeze_model)
+        model = decoder.pt_module
+        print(model.inlined_graph)
         im = fe.load(decoder)
         om = fe.convert(im)
         self._resolve_input_shape_dtype(om, ov_inputs, dynamic_shapes)
