+++ tests/layer_tests/pytorch_tests/test_add.py
@@ -111,7 +111,7 @@ class TestAddTypes(PytorchLayerTest):
         self.rhs_type = rhs_type
         self.rhs_shape = rhs_shape
         self._test(*self.create_model(lhs_type, lhs_shape, rhs_type, rhs_shape),
-                   ie_device, precision, ir_version)
+                   ie_device, precision, ir_version, freeze_model=False, trace_model=True)
 
 class TestAddLists(PytorchLayerTest):
